# backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.createMessageRequest:
    properties:
      content:
        type: string
    type: object
  internal_handler.errorResponse:
    properties:
      code:
        $ref: '#/definitions/perchly-backend_internal_apierror.Code'
      error:
        type: string
    type: object
...
  internal_handler.saveOnboardingProfileRequest:
    properties:
      age_range:
        type: string
      mood_preference:
        type: string
      notifications_granted:
        type: boolean
      preferred_name:
        description: |-
          PreferredName/SkipHitap are pointers deliberately: nil means the
          field was absent (an older client), which OnboardingService needs
          to tell apart from an explicit empty string / false — see its
          SaveProfile doc comment.
        type: string
      selected_persona_id:
        type: string
      skip_hitap:
        type: boolean
    type: object
...
  perchly-backend_internal_apierror.Code:
    enum:
    - invalid_request_body
    - unauthorized
    - internal_error
    - session_create_failed
    - invalid_email
    - email_code_send_failed
    - too_many_attempts
    - verification_code_expired
    - invalid_verification_code
    - verification_failed
    - invalid_display_name
    - user_not_found
    - profile_update_failed
    - invalid_refresh_token
    - session_refresh_failed
    - logout_failed
    - invalid_persona_id
    - persona_not_found
    - persona_fetch_failed
    - personas_list_failed
    - persona_not_age_appropriate
    - persona_trait_out_of_range
    - persona_traits_failed
    - invalid_conversation_id
    - conversation_not_found
    - conversation_forbidden
    - conversation_create_failed
    - conversations_list_failed
    - conversation_fetch_failed
    - conversation_context_missing
    - messages_list_failed
    - invalid_message_id
    - message_not_found
    - message_content_required
    - message_send_failed
    - streaming_unsupported
    - reaction_emoji_required
    - invalid_reaction_emoji
    - cannot_react_to_own_message
    - reaction_failed
    - invalid_age_range
    - invalid_mood_preference
    - invalid_selected_persona_id
    - preferred_name_required
    - preferred_name_invalid
    - onboarding_save_failed
    - onboarding_fetch_failed
    - onboarding_profile_not_found
    - quota_check_failed
    - quota_exceeded
    - chat_energy_failed
    type: string
    x-enum-varnames:
    - CodeInvalidRequestBody
    - CodeUnauthorized
    - CodeInternal
    - CodeSessionCreateFailed
    - CodeInvalidEmail
    - CodeEmailCodeSendFailed
    - CodeTooManyAttempts
    - CodeVerificationExpired
    - CodeInvalidCode
    - CodeVerificationFailed
    - CodeInvalidDisplayName
    - CodeUserNotFound
    - CodeProfileUpdateFailed
    - CodeInvalidRefreshToken
    - CodeSessionRefreshFailed
    - CodeLogoutFailed
    - CodeInvalidPersonaID
    - CodePersonaNotFound
    - CodePersonaFetchFailed
    - CodePersonasListFailed
    - CodePersonaNotAgeAppropriate
    - CodePersonaTraitOutOfRange
    - CodePersonaTraitsFailed
    - CodeInvalidConversationID
    - CodeConversationNotFound
    - CodeConversationForbidden
    - CodeConversationCreateFailed
    - CodeConversationsListFailed
    - CodeConversationFetchFailed
    - CodeConversationContextMissing
    - CodeMessagesListFailed
    - CodeInvalidMessageID
    - CodeMessageNotFound
    - CodeMessageContentRequired
    - CodeMessageSendFailed
    - CodeStreamingUnsupported
    - CodeReactionEmojiRequired
    - CodeInvalidReactionEmoji
    - CodeCannotReactToOwnMessage
    - CodeReactionFailed
    - CodeInvalidAgeRange
    - CodeInvalidMoodPreference
    - CodeInvalidSelectedPersonaID
    - CodePreferredNameRequired
    - CodePreferredNameInvalid
    - CodeOnboardingSaveFailed
    - CodeOnboardingFetchFailed
    - CodeOnboardingProfileNotFound
    - CodeQuotaCheckFailed
    - CodeQuotaExceeded
    - CodeChatEnergyFailed
...
  perchly-backend_internal_model.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      reaction_emoji:
        description: |-
          ReactionEmoji is the single emoji reaction left on this message by
          whichever side didn't send it (the user reacting to an assistant
          message, or the persona reacting to a user message — see
          ChatService.SendMessage), or nil if none. Always one of
          OrderedAllowedReactionEmojis.
        type: string
      role:
        $ref: '#/definitions/perchly-backend_internal_model.MessageRole'
    type: object
  perchly-backend_internal_model.MessageRole:
    enum:
    - user
    - assistant
    type: string
    x-enum-varnames:
    - MessageRoleUser
    - MessageRoleAssistant
  perchly-backend_internal_model.OnboardingProfile:
    properties:
      age_range:
        description: |-
          One of "under18", "age18to24", "age25to34", "age35plus" — matches
          the iOS OnboardingProfile.AgeRange enum's case names exactly.
        type: string
      created_at:
        type: string
      is_minor:
        description: |-
          Always recomputed server-side from AgeRange (see
          OnboardingService.SaveProfile) — never trusted verbatim from the
          client, since persona age-gating depends on it.
        type: boolean
      mood_preference:
        description: One of "motivation", "dailyChat", "hobbyTalk", "skipped", or
          nil.
        type: string
      notifications_granted:
        type: boolean
      preferred_name:
        description: |-
          PreferredName is exactly what the user typed for how they want to
          be addressed — never a server-generated nickname. nil means no
          hitap has been set (either they opted out via SkipHitap, or
          they're on a client version that doesn't send this yet).
        type: string
      selected_persona_id:
        type: string
      skip_hitap:
        description: |-
          SkipHitap is true when the user explicitly chose to stay
          anonymous — personas must never address them by any name,
          nickname, or invented handle in that case (see
          ContextBuilder.formatHitapInstruction).
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  perchly-backend_internal_model.Persona:
    properties:
      accent_color:
        type: string
      avatar_url:
        type: string
      category:
        type: string
      created_at:
        type: string
      default_traits:
        allOf:
        - $ref: '#/definitions/perchly-backend_internal_model.PersonaTraits'
        description: |-
          DefaultTraits are this persona's own personality dial positions —
          what a user gets before they've ever customized anything for it.
          See PersonaTraits and user_persona_traits.
      id:
        type: string
      is_active:
        type: boolean
      is_minor_appropriate:
        description: |-
          IsMinorAppropriate gates whether a user flagged as a minor (see
          OnboardingProfile.IsMinor) may start a conversation with this
          persona — enforced in ConversationService.Create.
        type: boolean
      name:
        type: string
      short_description:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      tone_description:
        type: string
      updated_at:
        type: string
    type: object
  perchly-backend_internal_model.PersonaTraits:
    properties:
      directness:
        description: |-
          Directness: how blunt/unsoftened vs. gentle and hedged its
          responses are — this is the dial behind "harsher answers".
        type: integer
      energy:
        description: 'Energy: how energetic/enthusiastic vs. calm and measured its
          tone is.'
        type: integer
      humor:
        description: 'Humor: how often and readily it jokes vs. stays fully serious.'
        type: integer
      warmth:
        description: 'Warmth: how affectionate/close vs. distant and formal the persona
          is.'
        type: integer
      wisdom:
        description: 'Wisdom: how deep/reflective vs. surface-level its responses
          are.'
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Perchly sohbet uygulamasının REST API'si — kimlik doğrulama, onboarding, persona, konuşma/mesajlaşma ve kota uçları.
    Kimlik doğrulama gerektiren uçlar için "Authorize" düğmesinden `Bearer <access_token>` girin.
  title: Perchly Backend API
  version: "1.0"
paths:
...
  /conversations/{id}/messages:
    get:
      description: Salt okunur; kota harcamaz. Sahiplik kontrolü sunucu tarafında
        yapılır.
      parameters:
      - description: Conversation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/perchly-backend_internal_model.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Konuşmadaki mesajları listele
      tags:
      - conversations
    post:
      consumes:
      - application/json
      description: 'Başarılı olursa cevap Server-Sent Events''e döner: her LLM chunk''ı
        için bir "message" event''i, persona kullanıcı mesajına tepki verdiyse tek
        bir "reaction" event''i, ve sonunda gerçek/persisted id''leri taşıyan bir
        "done" event''i (doneEventPayload); akış sırasında hata olursa "error" event''i
        gelir. Kota QuotaMiddleware tarafından bu uçtan önce kontrol edilir; günlük
        limit dolmuşsa ve kredi de yoksa 429 döner.'
      parameters:
      - description: Conversation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createMessageRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'text/event-stream: message/reaction/done/error event''leri'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "429":
          description: günlük kota doldu ve kredi yok
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Mesaj gönder (SSE stream)
      tags:
      - conversations
...
  /personas:
    get:
      description: Herkese açık liste. ?recommend=true ile (auth zorunlu) her personaya
        recommended/match_reason eklenmiş, kullanıcının mood_preference'ına göre kişiselleştirilmiş
        liste döner; tam olarak bir persona recommended:true olur, sayısal eşleşme
        % asla üretilmez.
      parameters:
      - description: true verilirse auth zorunlu, kişiselleştirilmiş öneri döner
        in: query
        name: recommend
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/perchly-backend_internal_model.Persona'
            type: array
        "401":
          description: recommend=true iken auth eksik
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Personaları listele
      tags:
      - personas
  /personas/{id}:
    get:
      parameters:
      - description: Persona ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/perchly-backend_internal_model.Persona'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Tek bir personayı getir
      tags:
      - personas
...
  /users/onboarding-profile:
    get:
      description: Yeniden kurulum veya ikinci bir cihazda daha önce verilen onboarding
        cevaplarını senkronize etmek için kullanılır.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/perchly-backend_internal_model.OnboardingProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: onboarding henüz tamamlanmamış
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Onboarding profilini getir
      tags:
      - onboarding
    post:
      consumes:
      - application/json
      description: 'Anonim kullanıcılar için de çalışır. is_minor asla client''tan
        okunmaz, age_range''den türetilir. preferred_name/skip_hitap opsiyoneldir:
        hiçbiri gönderilmezse eski istemci davranışı değişmez; skip_hitap=true ise
        preferred_name her zaman NULL''a zorlanır (sunucu asla takma ad üretmez);
        skip_hitap=false iken preferred_name boşsa 400 döner. Bilinmeyen alanlar yok
        sayılır.'
      parameters:
      - description: onboarding cevapları
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.saveOnboardingProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/perchly-backend_internal_model.OnboardingProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Onboarding profilini kaydet
      tags:
      - onboarding
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"